--- a/crates/gen/src/types/class.rs
+++ b/crates/gen/src/types/class.rs
@@ -178,7 +178,7 @@ impl Class {
                     };
                 }
                 unsafe impl ::winrt::RuntimeType for #name {
-                    const SIGNATURE: ::winrt::ConstString = { ::winrt::ConstString::from_slice(#signature) };
+                    const SIGNATURE: ::winrt::ConstBuffer = { ::winrt::ConstBuffer::from_slice(#signature) };
                 }
                 unsafe impl ::winrt::AbiTransferable for #name {
                     type Abi = ::winrt::RawComPtr<#default_name>;