--- a/crates/gen/src/types/delegate.rs
+++ b/crates/gen/src/types/delegate.rs
@@ -109,9 +109,9 @@ impl Delegate {
             }
             unsafe impl<#constraints> ::winrt::ComInterface for #name {
                 type VTable = #abi_definition;
-                fn iid() -> ::winrt::Guid {
+                const IID: ::winrt::Guid = {
                     #guid
-                }
+                };
             }
             impl<#constraints> ::std::clone::Clone for #name {
                 fn clone(&self) -> Self {
@@ -130,9 +130,7 @@ impl Delegate {
                 #phantoms
             }
             unsafe impl<#constraints> ::winrt::RuntimeType for #name {
-                fn signature() -> String {
-                    #signature
-                }
+                const SIGNATURE: &'static [u8] = { #signature };
             }
             unsafe impl<#constraints> ::winrt::AbiTransferable for #name {
                 type Abi = ::winrt::RawComPtr<Self>;
@@ -192,9 +190,9 @@ impl Delegate {
                     unsafe {
                         let this: *mut Self = this.as_raw() as _;
 
-                        if iid == &<#name as ::winrt::ComInterface>::iid()
-                            || iid == &<::winrt::IUnknown as ::winrt::ComInterface>::iid()
-                            || iid == &<::winrt::IAgileObject as ::winrt::ComInterface>::iid()
+                        if iid == &<#name as ::winrt::ComInterface>::IID
+                            || iid == &<::winrt::IUnknown as ::winrt::ComInterface>::IID
+                            || iid == &<::winrt::IAgileObject as ::winrt::ComInterface>::IID
                         {
                             *interface = this as ::winrt::RawPtr;
                             (*this).count.add_ref();