--- a/crates/gen/src/types/delegate.rs
+++ b/crates/gen/src/types/delegate.rs
@@ -130,7 +130,7 @@ impl Delegate {
                 #phantoms
             }
             unsafe impl<#constraints> ::winrt::RuntimeType for #name {
-                const SIGNATURE: ::winrt::ConstString = { #signature };
+                const SIGNATURE: ::winrt::ConstBuffer = { #signature };
             }
             unsafe impl<#constraints> ::winrt::AbiTransferable for #name {
                 type Abi = ::winrt::RawComPtr<Self>;