--- a/crates/gen/src/types/enum.rs
+++ b/crates/gen/src/types/enum.rs
@@ -85,7 +85,7 @@ impl Enum {
                 #(#fields)*
             }
             unsafe impl ::winrt::RuntimeType for #name {
-                const SIGNATURE: &'static [u8] = { #signature };
+                const SIGNATURE: ::winrt::ConstString = { ::winrt::ConstString::from_slice(#signature) };
             }
             unsafe impl ::winrt::AbiTransferable for #name {
                 type Abi = #repr;