--- a/crates/gen/src/types/interface.rs
+++ b/crates/gen/src/types/interface.rs
@@ -86,7 +86,7 @@ impl Interface {
             }
             unsafe impl<#constraints> ::winrt::ComInterface for #name {
                 type VTable = #abi_definition;
-                const IID: ::winrt::Guid = { #guid };
+                const IID: ::winrt::Guid = #guid;
             }
             impl<#constraints> ::std::clone::Clone for #name {
                 fn clone(&self) -> Self {