--- a/src/activation_factory.rs
+++ b/src/activation_factory.rs
@@ -25,14 +25,14 @@ impl IActivationFactory {
 unsafe impl ComInterface for IActivationFactory {
     type VTable = abi_IActivationFactory;
 
-    fn iid() -> Guid {
+    const IID: Guid = {
         Guid::from_values(
             0x0000_0035,
             0x0000,
             0x0000,
             [0xc0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x46],
         )
-    }
+    };
 }
 
 unsafe impl AbiTransferable for IActivationFactory {