… .github/workflows/build.yml
… Cargo.toml
… src/com/interface.rs
… src/com/ptr.rs
… src/com/raw_ptr.rs
… src/com/try_into.rs
… src/com/unknown.rs
… src/error.rs
… src/factory.rs
… src/guid.rs
… src/lib.rs
… src/object.rs
… src/runtime_type.rs
… src/time_span.rs
… tests/delegates.rs
… tests/generic_guids.rs
… tests/interface.rs
… tests/uri.rs