Fixes #136 

Note that this will push the MSRV to Rust 1.46 which is currently in beta.