@rylev 2020-07-22
@kennykerr this is ready for review. It won't pass CI until I update const-sha1 with the changes necessary to get this to work, but I'd like to have your eyes on this before I push a new version. Once that's done, I'll push a new version of const-sha1 and we can merge this.