@kennykerr 2020-07-22
Can't wait to do some run time comparative performance testing with this change.