@Boddlnagg 2020-08-03
Note that this is a breaking change for consumers of `winrt-rs` who have their own impls of `winrt::ComInterface`. This should be reflected in the next release version.