@rylev 2020-08-03
Yes, our next version will be 0.8 for multiple reasons 👍 .