@kennykerr 2020-08-04
Here's the performance impact of const guids. Consider the following sample:

```rust
fn main() -> Result<()> {
    let uri = Uri::create_uri("http://kennykerr.ca")?;
    let query = uri.query_parsed()?;

    let start = std::time::Instant::now();

    for _ in 0..1_000_000 {
        let s = query.size()?;
        debug_assert!(s == 0);
    }

    println!("{:5}ms", start.elapsed().as_millis());
    Ok(())
}
```

The `size` method is on a required but non-default interface of the `WwwFormUrlDecoder` class returned by the `query_parsed` method. That means that every time `size` is called, `QueryInterface` must first be called to retrieve the appropriate vtable. This is just one of those details that a language projection hides from the language as a convenience, but such hidden costs need to be as minimal as possible.

Before this commit:

```
C:\git\rust\scratch>cargo run --release
    Finished release [optimized] target(s) in 0.09s
     Running `target\release\scratch.exe`
  829ms
```

After this commit:

```
C:\git\rust\scratch>cargo run --release
    Finished release [optimized] target(s) in 0.09s
     Running `target\release\scratch.exe`
   21ms
```

That's about 40 times faster and on par with C++/WinRT.