@nu8 2020-08-27
@kennykerr looks like your snippet is missing this:

~~~rust
import!(
   dependencies
      os
   types
      windows::foundation::Uri
);
use windows::foundation::Uri;
use winrt::*;
~~~

otherwise its not valid Rust.